+++ flowgraph.hpp
@@ -49,7 +49,7 @@ struct FlowGraph
 	void build();
 	void calculate();
 	void dump(std::string name) const;
-	int edge_from(const Edge* edge) const;
-	int edge_to(const Edge* edge) const;
+	size_t edge_from(const Edge* edge) const;
+	size_t edge_to(const Edge* edge) const;
 };
